% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{acronym}
\alias{acronym}
\title{Generate acronym from input string}
\usage{
acronym(
  input,
  dictionary = NULL,
  acronym_length = 3,
  ignore_articles = TRUE,
  timeout = 60,
  bow = FALSE,
  bow_prop = 0.5,
  to_tibble = FALSE
)
}
\arguments{
\item{input}{Character vector with text to use as the input for the candidate}

\item{dictionary}{Character vector containing dictionary of terms from which acronym should be created; default is \code{NULL} and \code{hunspell} "en_us" dictionary will be used}

\item{acronym_length}{Number of characters in acronym; default is \code{3}}

\item{ignore_articles}{Boolean indicating whether or not articles should be ignored ; default is \code{TRUE}}

\item{timeout}{Number of seconds to spend trying to search the for an acronym; default is \code{60}}

\item{bow}{Boolean for whether or not a "bag of words" approach should be used for "input" vector; default is \code{FALSE}}

\item{bow_prop}{Given \code{bow = TRUE} this specifies the proportion of words to sample; ignored if \code{bow = FALSE}; default is \code{0.5}}

\item{to_tibble}{Boolean as to whether or not the result should be a \code{tibble}; default is \code{FALSE}}
}
\value{
If \code{to_tibble = FALSE} (default), then a character vector containing the acronym capitalized and the original string, with letters used in the acronym capitalized.

If \code{to_tibble = TRUE}, then a \code{tibble} with the following columns:
\itemize{
\item \strong{formatted}: The candidate acronym and string with letters used capitalized
\item \strong{prefix}: The candidate acronym
\item \strong{suffix}: Words used with letters in acronym capitalized
\item \strong{original}: The original string used to construct the acronym
}
}
\description{
Generate acronym from input string
}
