% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_articles}
\alias{find_articles}
\title{Helper to find articles}
\usage{
find_articles(word)
}
\arguments{
\item{word}{Character vector of length 1 with word to check}
}
\value{
Logical vector of length one, \code{TRUE} if the word is an article and \code{FALSE} if not.
}
\description{
This function will check if an input word is an article in the English language ('a', 'an', 'the').
}
\examples{

find_articles("the")
find_articles("then")
find_articles("whatever")

}
