% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{initialism}
\alias{initialism}
\title{Generate initialism from input string}
\usage{
initialism(
  input,
  ignore_articles = TRUE,
  bow = FALSE,
  bow_prop = 0.5,
  to_tibble = FALSE
)
}
\arguments{
\item{input}{Character vector with text to use as the input for the candidate}

\item{ignore_articles}{Boolean indicating whether or not articles should be ignored ; default is \code{TRUE}}

\item{bow}{Boolean for whether or not a "bag of words" approach should be used for "input" vector; default is \code{FALSE}}

\item{bow_prop}{Given \code{bow = TRUE} this specifies the proportion of words to sample; ignored if \code{bow = FALSE}; default is \code{0.5}}

\item{to_tibble}{Boolean as to whether or not the result should be a \code{tibble}; default is \code{FALSE}}
}
\value{
If \code{to_tibble = FALSE} (default), then a character vector containing the initialism capitalized and the original string, with letters used in the initialism capitalized.

If \code{to_tibble = TRUE}, then a \code{tibble} with the following columns:
\itemize{
\item \strong{formatted}: The initialism and string with letters used capitalized
\item \strong{prefix}: The initialism
\item \strong{suffix}: Words used with letters in initialism capitalized
\item \strong{original}: The original string used to construct the initialism
}
}
\description{
Generate initialism from input string
}
