% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_input}
\alias{process_input}
\title{Process the input string}
\usage{
process_input(input, ignore_articles = TRUE, bow = FALSE, bow_prop = 0.5)
}
\arguments{
\item{input}{Character vector with text to use as the input for the candidate}

\item{ignore_articles}{Boolean indicating whether or not articles should be ignored ; default is `TRUE`}

\item{bow}{Boolean for whether or not a "bag of words" approach should be used for "input" vector; default is `FALSE`}

\item{bow_prop}{Given `bow = TRUE` this specifies the proportion of words to sample; ignored if `bow = FALSE`; default is `0.5`}
}
\value{
Named list with the following elements:

- **words**: Vector with one element per word to be used in the acronym or initialism
- **collapsed**: Vector of length 1 containing all characters from words collapsed
- **words_len**: Vector containing lengths of each word
- **first_chars**: Vector containing first character from each word
}
\description{
This helper is used by both \link[acroname]{acronym} and \link{initialism} to extract elements needed from the input string.

If the function is used with `bow = TRUE` the input will be processed with a "bag of words" approach, by which words will be shuffled and sampled without replacement. In thi used will be determined by the proportion passed to "bow_prop".
}
